% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_helpers.R
\name{clps}
\alias{clps}
\title{Wrapper Around Paste to Collapse a Character Vector}
\usage{
clps(collapse = " ", ..., sep = " ")
}
\arguments{
\item{collapse}{an optional character string to separate the results.  Not
    \code{\link{NA_character_}}.}

\item{...}{one or more \R objects, to be converted to character vectors.}

\item{sep}{a character string to separate the terms.  Not
    \code{\link{NA_character_}}.}

\item{...}{Arguments passed on to \code{base::paste}
\describe{
  \item{sep}{a character string to separate the terms.  Not
    \code{\link{NA_character_}}.}
  \item{collapse}{an optional character string to separate the results.  Not
    \code{\link{NA_character_}}.}
}}
}
\description{
Wrapper Around Paste to Collapse a Character Vector
}
\keyword{internal}
