% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-mirt.R
\name{control_mirt}
\alias{control_mirt}
\title{Control aspects of fitting a model in mirt}
\usage{
control_mirt(
  SE = TRUE,
  quadpts = NULL,
  control = list(),
  technical = list(NCYCLES = NULL),
  ...
)
}
\arguments{
\item{SE}{logical; estimate the standard errors by computing the parameter information matrix?
See \code{SE.type} for the type of estimates available}

\item{quadpts}{number of quadrature points per dimension (must be larger than 2).
By default the number of quadrature uses the following scheme:
\code{switch(as.character(nfact), '1'=61, '2'=31, '3'=15, '4'=9, '5'=7, 3)}.
However, if the method input is set to \code{'QMCEM'} and this argument is left blank then
the default number of quasi-Monte Carlo integration nodes will be set to 5000 in total}

\item{control}{List of arguments passed to argument \code{control} of
\code{\link[mirt:mirt]{mirt::mirt()}}. See examples below.}

\item{technical}{List of arguments passed to argument \code{technical} of
\code{\link[mirt:mirt]{mirt::mirt()}}. See examples below.}

\item{...}{Other arguments passed to \code{\link[mirt:mirt]{mirt::mirt()}}.}
}
\value{
A list with one element for every argument of \code{control_mirt()}.
}
\description{
This function should be used to generate the \code{control} argument of the
\code{\link[=fit.irtree_model]{fit()}} function.
}
\examples{
control_mirt(SE = FALSE,
             quadpts = 15,
             control = list(),
             technical = list(NCYCLES = 500),
             TOL = .0001)
}
