% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-mirt-output.R
\name{extract_mirt_output-deprecated}
\alias{extract_mirt_output-deprecated}
\title{Retrieve estimates from mirt.}
\usage{
extract_mirt_output(results = NULL, object = NULL, method = "MAP",
  class = NULL, ...)
}
\arguments{
\item{results}{An object of class \code{\link[mirt]{SingleGroupClass-class}}
as returned from \code{\link[=irtree_fit_mirt]{irtree_fit_mirt()}}.}

\item{object}{A description of the user-specified model. See
\link{irtree_model} for more information.}

\item{method}{Passed to \code{\link[mirt:fscores]{mirt::fscores()}}.}

\item{class}{String specifying which class of model was fit.}

\item{...}{Passed to \code{\link[mirt:fscores]{mirt::fscores()}}.}
}
\value{
A list of parameter estimates and model fit information.
}
\description{
This function takes the output from \code{\link[=irtree_fit_mirt]{irtree_fit_mirt()}} and
returns the parameter estimates in a convenient way.
}
\seealso{
\code{\link{ItemResponseTrees-deprecated}}
}
\keyword{internal}
