% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_fit_mirt.R
\name{extract_mirt_output}
\alias{extract_mirt_output}
\title{Retrieve Estimates From mirt}
\usage{
extract_mirt_output(results = NULL, model = NULL, method = "MAP",
  class = NULL, ...)
}
\arguments{
\item{results}{An object of class \code{\link[mirt]{SingleGroupClass-class}}
as returned from \code{\link{fit_tree_mirt}}.}

\item{model}{A description of the user-specified model. See
\code{\link{tree_model}} for more information.}

\item{method}{type of factor score estimation method. Can be:
\itemize{
    \item \code{"EAP"} for the expected a-posteriori (default)
    \item \code{"MAP"} for the maximum a-posteriori (i.e, Bayes modal)
    \item \code{"ML"} for maximum likelihood
    \item \code{"WLE"} for weighted likelihood estimation
    \item \code{"EAPsum"} for the expected a-posteriori for each sum score
    \item \code{"plausible"} for a single plausible value imputation for each case.
      This is equivalent to setting \code{plausible.draws = 1}
    \item \code{"classify"} for the posteriori classification probabilities (only
      applicable when the input model was of class \code{MixtureClass})
 }}

\item{class}{String specifying which class of model was fit}

\item{...}{Further arguments passed to \code{\link[mirt]{fscores}}}
}
\value{
A list of parameter estimates and model fit information.
}
\description{
This function takes the output from \code{\link[mirt]{mirt}} and returns the
parameter estimates in a convenient way.
}
