% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-mplus-output.R
\name{extract_mplus_output-deprecated}
\alias{extract_mplus_output-deprecated}
\title{Retrieve estimates From Mplus.}
\usage{
extract_mplus_output(results = NULL, object = NULL, class = NULL,
  .errors2messages = FALSE)
}
\arguments{
\item{results}{A list as returned from \code{\link[=irtree_fit_mplus]{irtree_fit_mplus()}}.}

\item{object}{A description of the user-specified model. See
\link{irtree_model} for more information.}

\item{class}{String specifying which class of model was fit}

\item{.errors2messages}{Logical indicating whether errors should be converted
to messages}
}
\value{
A list of parameter estimates, model fit information
(\code{summaries}), \code{warnings}, \code{errors}.
}
\description{
This function takes the output from Mplus as returned from
\code{\link[=irtree_fit_mplus]{irtree_fit_mplus()}} and returns the estimates in a convenient way.
}
\seealso{
\code{\link{ItemResponseTrees-deprecated}}
}
\keyword{internal}
