% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-mplus.R
\name{extract_mplus_output}
\alias{extract_mplus_output}
\title{Retrieve Estimates From Mplus}
\usage{
extract_mplus_output(results = NULL, object = NULL, class = NULL,
  .errors2messages = FALSE)
}
\arguments{
\item{results}{A list as returned from \code{\link[MplusAutomation]{readModels}}}

\item{object}{A description of the user-specified model. See
\link{irtree_model} for more information.}

\item{class}{String specifying which class of model was fit}

\item{.errors2messages}{Logical indicating whether errors should be converted
to messages}
}
\value{
A list of parameter estimates, model fit information
(\code{summaries}), \code{warnings}, \code{errors}.
}
\description{
This function takes the output from Mplus as returned from
\code{\link[MplusAutomation]{readModels}} and returns the estimates in a
convenient way.
}
