% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit.irtree_model}
\alias{fit.irtree_model}
\title{Fit an IR-Tree model.}
\usage{
\method{fit}{irtree_model}(
  object = NULL,
  data = NULL,
  engine = c("mplus", "mirt", "tam"),
  ...,
  link = c("logit", "probit"),
  verbose = interactive(),
  control = NULL,
  improper_okay = FALSE
)
}
\arguments{
\item{object}{Object of class \code{irtree_model}. See \link{irtree_model} for more
information.}

\item{data}{Data frame containing containing one row per respondent and one
column per variable. The variable names must correspond to those used in
\code{object}.}

\item{engine}{String specifying whether to use mirt, Mplus, or TAM for
estimation.}

\item{...}{Not currently used. Use \code{control} instead.}

\item{link}{String specifying the link function. May be either logit, or (in
case of Mplus), probit.}

\item{verbose}{Logical indicating whether output should be printed to the
console.}

\item{control}{List. The allowed elements of this list depend on the
\code{engine}. Use \link{control_mirt}, \link{control_mplus}, or \link{control_tam} for
convenience. Note that the \code{fit()} function does not use \code{...}, but that you
can use the \verb{control_*()} functions to pass additional arguments.}

\item{improper_okay}{Logical indicating whether the model should also be fit
if it is not a proper IR-tree model. Set this only to \code{TRUE} if you really
know what you are doing.}
}
\value{
Returns a list of class \code{irtree_fit}. The first list element is the
return value of either \code{\link[mirt:mirt]{mirt::mirt()}}, \code{\link[MplusAutomation:readModels]{MplusAutomation::readModels()}} or
\code{\link[TAM:tam.mml]{TAM::tam.mml()}}. Further information is provided in the element
\code{spec}.
}
\description{
This function takes a \code{data} frame and an \code{object} of class \link{irtree_model}
and runs the model in either mirt, Mplus, or TAM.
}
\section{Methods}{
 The methods \code{coef()}, \code{summary()}, and \code{print()} are
implemented for objects of class \code{irtree_fit}, and those wrap the
respective functions of \link[mirt:mirt-package]{mirt},
\link[MplusAutomation:MplusAutomation-package]{MplusAutomation}, or
\link[TAM:TAM-package]{TAM}. However, \code{\link[=glance.irtree_fit]{glance()}},
\code{\link[=tidy.irtree_fit]{tidy()}}, and \code{\link[=augment.irtree_fit]{augment()}} may be
more helpful.
}

\examples{
m1 <- "
IRT:
attitude BY Comfort, Work, Future, Benefit;
Class:
GRM
"

model1 <- irtree_model(m1)

data(Science, package = "mirt")

fit1 <- fit(model1, data = Science, engine = "mirt",
            control = control_mirt(quadpts = 15))

glance(fit1)
tidy(fit1)
augment(fit1)
}
