% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_fit_mirt.R
\name{fit_tree_mirt}
\alias{fit_tree_mirt}
\title{Fit an IR-Tree Model using mirt}
\usage{
fit_tree_mirt(data = NULL, model = NULL, dir = ".", link = "logit",
  SE = TRUE, verbose = FALSE, run = TRUE, rm_mirt_internal = TRUE,
  ...)
}
\arguments{
\item{data}{Data frame containing only the items.}

\item{model}{A description of the user-specified model. See
\code{\link{tree_model}} for more information.}

\item{dir}{String, path name where the Mplus files should be
stored and run; the default corresponds to the working directory}

\item{link}{String specifying the link function. Only \code{logit} is
implemented in mirt.}

\item{SE}{logical; estimate the standard errors by computing the parameter information matrix?
See \code{SE.type} for the type of estimates available}

\item{verbose}{logical; print observed- (EM) or complete-data (MHRM) log-likelihood
after each iteration cycle? Default is TRUE}

\item{run}{Logical, whether to indeed run Mplus.}

\item{rm_mirt_internal}{Logical. \code{\link[mirt]{mirt}} returns a lot of
information including two functions that can take up a huge amount of space
(https://github.com/philchalmers/mirt/issues/147#issue-352032654). These
two functions are removed from the output if \code{rm_mirt_internal =
  TRUE}.
... Other arguments passed to \code{\link[mirt]{mirt}}.}

\item{...}{additional arguments to be passed}
}
\value{
List with two elements. \code{mirt} contains the mirt output, namely
an object of class \code{\link[mirt]{SingleGroupClass-class}} . \code{args}
contains the input specifications.
}
\description{
This function takes a data frame and a model string and runs the model in mirt.
}
