% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_fit_mplus.R
\name{fit_tree_mplus}
\alias{fit_tree_mplus}
\title{Fit an IR-Tree Model using Mplus}
\usage{
fit_tree_mplus(data = NULL, model = NULL, dir = ".",
  file_name = "mplus-tree", integration_points = 15,
  estimator = "MLR", link = c("probit", "logit"), inp_rm = FALSE,
  out_rm = FALSE, fsc_rm = FALSE, dat_rm = FALSE,
  save_fscores = TRUE, analysis_list = list(), run = TRUE,
  showOutput = FALSE, replaceOutfile = "always", overwrite = FALSE,
  .warnings2messages = FALSE)
}
\arguments{
\item{data}{Data frame containing only the items.}

\item{model}{A description of the user-specified model. See
\code{\link{tree_model}} for more information.}

\item{dir}{Character string of path name, where the Mplus files should be
stored and run; the default corresponds to the working directory.}

\item{file_name}{Character string, file name of the data file and the Mplus files.}

\item{integration_points}{Integer, passed to argument 'INTEGRATION' in Mplus.
Number of integration points for numerical integration.}

\item{estimator}{Character string, passed to argument 'ESTIMATOR' in Mplus.}

\item{link}{Character string, passed to argument 'LINK' in Mplus. Specifies
the link function.}

\item{inp_rm}{Logical, whether the Mplus input file should be removed on exit.}

\item{out_rm}{Logical, whether the Mplus output file should be removed on exit.}

\item{fsc_rm}{Logical, whether the Mplus FSCORES file should be removed on exit.}

\item{dat_rm}{Logical, whether the data file should be removed on exit.}

\item{save_fscores}{Logical, wheter to save FSCORES or not.}

\item{analysis_list}{Named list of strings passed to Mplus' argument
ANALYSIS. For example: \code{analysis_list = list(MITERATIONS = "1000")}.}

\item{run}{Logical, whether to indeed run Mplus.}

\item{showOutput}{optional. If \code{TRUE}, show estimation output (TECH8)
in the R console. Note that if run within Rgui, output will display within R,
but if run via Rterm, a separate window will appear during estimation.}

\item{replaceOutfile}{optional. Currently supports three settings: \dQuote{always}, which
runs all models, regardless of whether an output file for the model exists; \dQuote{never},
which does not run any model that has an existing output file; and \dQuote{modifiedDate}, which
only runs a model if the modified date for the input file is more recent than
the output file modified date (implying there have been updates to the model).}

\item{overwrite}{Logical value indicating whether
data and input (if present) files should be overwritten.
Defaults to \code{TRUE} to be consistent with prior behavior.
If \code{FALSE} and the file to write the data to already exists,
it will throw an error.}

\item{.warnings2messages}{Logial, whether Mplus errors and warnings should be
signaled as warnings (the default) or messages.}
}
\value{
List with two elements. \code{Mplus} contains the Mplus output read into R via \code{\link[MplusAutomation]{readModels}}. \code{args} contains the input specifications.
}
\description{
This function takes a data frame and a model string and runs the model in Mplus.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
