% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-mplus.R
\name{irtree_fit_mplus}
\alias{irtree_fit_mplus}
\title{Fit an IR-Tree Model using Mplus}
\usage{
irtree_fit_mplus(
  object = NULL,
  data = NULL,
  file = tempfile("irtree_"),
  quadpts = 15,
  estimator = "MLR",
  link = c("probit", "logit"),
  run = TRUE,
  cleanup = run,
  save_fscores = TRUE,
  analysis_list = list(COVERAGE = "0"),
  verbose = interactive(),
  replaceOutfile = "always",
  overwrite = FALSE,
  ...,
  .warnings2messages = FALSE,
  .improper_okay = FALSE
)
}
\arguments{
\item{object}{Object of class \code{irtree_model}. See \link{irtree_model} for more
information.}

\item{data}{Data frame containing containing one row per respondent and one
column per variable. The variable names must correspond to those used in
\code{object}.}

\item{file}{String naming the file (or path) of the Mplus files and the data
file. Do not provide file endings, those will be automatically appended by
the function.}

\item{quadpts}{This is passed to argument 'INTEGRATION'. Thus, it may be an
integer specifiying the number of integration points for the Mplus default
of rectangular numerical integration (e.g., \code{quadpts = 15}). Or it may be a
string, which gives more fine grained control (e.g., \code{quadpts = "MONTECARLO(2000)"}).}

\item{estimator}{String, passed to argument 'ESTIMATOR' in Mplus.}

\item{link}{String, passed to argument 'LINK' in Mplus. Specifies
the link function.}

\item{run}{Logical, whether to indeed run Mplus.}

\item{cleanup}{Logical, whether the Mplus files should be removed on exit.}

\item{save_fscores}{Logical, wheter to save FSCORES or not.}

\item{analysis_list}{Named list of strings passed to Mplus' argument
ANALYSIS. For example: \code{analysis_list = list(MITERATIONS = "1000")}.}

\item{verbose}{Logical indicating whether Mplus/mirt output should be printed
to the console.}

\item{replaceOutfile}{optional. Currently supports three settings: \dQuote{always}, which
runs all models, regardless of whether an output file for the model exists; \dQuote{never},
which does not run any model that has an existing output file; and \dQuote{modifiedDate}, which
only runs a model if the modified date for the input file is more recent than
the output file modified date (implying there have been updates to the model).}

\item{overwrite}{Logical value indicating whether
data and input (if present) files should be overwritten.
Defaults to \code{TRUE} to be consistent with prior behavior.
If \code{FALSE} and the file to write the data to already exists,
it will throw an error.}

\item{...}{Additional parameters passed to \code{\link[MplusAutomation:runModels]{MplusAutomation::runModels()}}.}

\item{.warnings2messages}{Logial, whether Mplus errors and warnings should be
signaled as warnings (the default) or messages.}
}
\value{
List with two elements. \code{mplus} contains the Mplus output read into R
via \code{\link[MplusAutomation:readModels]{MplusAutomation::readModels()}}. \code{spec} contains the input
specifications.
}
\description{
This function takes a data frame and a model string and runs the model in Mplus.
}
\examples{
\donttest{
m1 <- "
IRT:
attitude BY Comfort, Work, Future, Benefit;

Class:
GRM
"
model1 <- irtree_model(m1)
data(Science, package = "mirt")

fit1 <- fit(model1, Science, engine = "mplus")
}
}
