% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{irtree_sim}
\alias{irtree_sim}
\title{Run a Simulation by Generating From and Fitting a Tree Model}
\usage{
irtree_sim(gen_model = NULL, fit_model = gen_model, N = NULL,
  sigma = NULL, itempar = NULL, link = c("probit", "logit"),
  engine = c("mirt", "mplus"), verbose = FALSE, save_rdata = TRUE,
  file = NULL, R = 1, plan = NULL, future_args = list(),
  in_memory = c("reduced", "everything", "nothing"), dots = list(),
  .dir = tempdir(), .na_okay = TRUE)
}
\arguments{
\item{gen_model}{Object of class \code{irtree_model} describing the
data-generating model. See \link{irtree_model} for more information.}

\item{fit_model}{Object of class \code{irtree_model} describing the model that
should be fit to the data. May be a list of multiple objects of class
\code{irtree_model} if different models should be fit to the same data set. See
\link{irtree_model} for more information.}

\item{N}{Integer, the number of persons.}

\item{sigma}{Either a matrix or a function that returns a matrix. This matrix
is the variance-covariance matrix of the person parameters that is passed
to \code{\link[MASS:mvrnorm]{MASS::mvrnorm()}}. Note that the order of the person
parameters is taken from the section Processes in the model \code{object} (see
\code{\link[=irtree_model]{irtree_model()}}).}

\item{itempar}{Either a list or a function that returns a list. The list has
an element \code{beta} and an element \code{alpha}. Each of these is a
matrix of item parameters. Note that the order of items (rows) is taken from the
section Items and the order of processes (columns) is taken from the
section Processes in the \code{model} (see \code{\link{irtree_model}}).}

\item{link}{Character. Link function to use.}

\item{engine}{String specifying whether to use Mplus or mirt for estimation.}

\item{verbose}{Logical indicating whether Mplus/mirt output should be printed
to the console.}

\item{save_rdata}{Logical indicating whether to save the results to an RData
file.}

\item{file}{String giving the file path used to save the output if
\code{save_rdata = TRUE}. Note that the file ending is automatically set to
\code{.rda}. This argument is also passed to \code{\link[=irtree_fit_mplus]{irtree_fit_mplus()}} if applicable.}

\item{R}{Number of replications. Can be either a single number indicating the
number of replications (e.g., \code{R = 100}), or can be a range (e.g., \code{R = 1:100}).}

\item{plan}{Parameter passed as argument \code{strategy} to \code{\link[future:plan]{future::plan()}}. May
be set to, for example, \code{multisession} in order to run the simulations in
parallel.}

\item{future_args}{Named list. Parameters passed \code{\link[future:plan]{future::plan()}}.}

\item{in_memory}{Character string indicating what output should be kept in
memory (note the argument \code{save_rdata}, which is not affected by
\code{in_memory}).}

\item{dots}{Nested list used to pass further arguments to downstream
methods/functions (i.e., \code{\link[=fit.irtree_model]{fit()}},
\code{\link[=tidy.irtree_fit]{tidy()}}). This is a named list, where the name of
each top-level element corresponds to the respective function. Each
top-level element is again a list, where the names of each element
correspond to the arguments. For example: \code{dots = list(fit = list(SE = FALSE), tidy = list(difficulty = TRUE))}.}

\item{.dir}{Path name that is used to save the results of every run if
\code{save_rdata = TRUE}.}

\item{.na_okay}{Logical indicating whether variables with unobserved response
categories are permitted. If not (\code{.na_okay = FALSE}), rejection sampling
is used to ensure that all categories are observed.}

\item{...}{Other parameters passed \code{\link[=irtree_sim1]{irtree_sim1()}}.}
}
\description{
The function \code{\link[=irtree_sim1]{irtree_sim1()}} generates a data set from an IR-tree
model and fits one or more models to that data set. Herein, this
process is repeated \code{R} times, and the argument \code{plan} allows to
run the simulation in parallel.
}
\seealso{
\code{\link[=irtree_sim1]{irtree_sim1()}}
