% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-data.R
\name{irtree_sim_data}
\alias{irtree_sim_data}
\title{Generate Data From an IR-Tree Model}
\usage{
irtree_sim_data(object = NULL, N = NULL, sigma = NULL,
  itempar = NULL, link = c("probit", "logit"), K = NULL)
}
\arguments{
\item{object}{A description of the user-specified model. See
\link{irtree_model} for more information.}

\item{N}{Integer, the number of persons.}

\item{sigma}{Either a matrix or a function that returns a matrix. This matrix
is the variance-covariance matrix of the person parameters that is passed
to \code{\link[MASS:mvrnorm]{MASS::mvrnorm()}}. Note that the order of the person
parameters is taken from the section Processes in the model \code{object} (see
\code{\link[=irtree_model]{irtree_model()}}).}

\item{itempar}{Either a list or a function that returns a list. The list has
an element \code{beta} and an element \code{alpha}. Each of these is a
matrix of item parameters. Note theat the order of items (rows) is taken from the
section Items and the order of processes (columns) is taken from the
section Processes in the \code{model} (see \code{\link{irtree_model}}).}

\item{link}{Character. Link function to use.}

\item{K}{Integer, number of categories. Needed only not defined by the
equations in \code{model}.}
}
\value{
A list with element \code{X} containing the data and an
element \code{args} containing the true parameter values etc.
}
\description{
This function generates data from an IR-tree model.
}
