% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ItemResponseTrees-deprecated.R,
%   R/extract-mirt-output.R, R/extract-mplus-output.R
\name{ItemResponseTrees-deprecated}
\alias{ItemResponseTrees-deprecated}
\alias{extract_mirt_output}
\alias{extract_mplus_output}
\title{Deprecated functions in package \strong{ItemResponseTrees}}
\description{
These functions are provided for compatibility with older versions only,
and may be defunct as soon as the next release.
}
\details{
The original help page for these functions is available at
help("oldName-deprecated") (note the quotes).
}
\section{\code{extract_mirt_output()}}{

This function is deprecated. Use \code{glance()}, \code{tidy()}, and \code{augment()}
instead.
}

\section{\code{extract_mplus_output()}}{

This function is deprecated. Use \code{glance()}, \code{tidy()}, and \code{augment()}
instead.
}

\keyword{internal}
