% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{myTryCatch}
\alias{myTryCatch}
\title{Custom \code{tryCatch()} that also returns the value of the expression}
\usage{
myTryCatch(expr)
}
\arguments{
\item{expr}{expression to be evaluated}
}
\value{
list with three elements, namely, \code{value} giving the value of
\code{expr}, \code{warning} giving any warning messages, and \code{error}
giving any error messages.
}
\description{
Custom \code{tryCatch()} that also returns the value of the expression
}
\details{
This function is based on an
\href{https://stackoverflow.com/a/24569739}{answer} by
\href{https://stackoverflow.com/users/2161065}{user2161065} on Stack Overflow.
User contributions on Stack Overflow are licensed under \href{https://creativecommons.org/licenses/by-sa/4.0/}{CC BY-SA 4.0}.
}
\examples{
ItemResponseTrees:::myTryCatch(log(1))
ItemResponseTrees:::myTryCatch(log(-1))
ItemResponseTrees:::myTryCatch(log("a"))
}
\seealso{
\code{\link[=tryCatch]{tryCatch()}}
}
\keyword{internal}
