% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_gen_tree_data.R
\name{recode_data}
\alias{recode_data}
\title{Recode Data Set Into Pseudoitems}
\usage{
recode_data(model = NULL, data = NULL, keep = FALSE)
}
\arguments{
\item{model}{A description of the user-specified model. See
\code{\link{tree_model}} for more information.}

\item{data}{Data frame containing only the items.}

\item{keep}{Logical indicating whether to append the original items to the
data frame of the generated pseudoitems}
}
\value{
Data frame
}
\description{
This function takes a data set with polytomous items and an IR-tree model and
returns the recoded items, the so-called pseudoitems.
}
