% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rtruncatednorm}
\alias{rtruncatednorm}
\title{Sample from a truncated normal distribution}
\usage{
rtruncatednorm(n = NULL, mean = 0, sd = 1, ll = -Inf, ul = Inf)
}
\arguments{
\item{n}{Number of observations}

\item{mean}{Mean}

\item{sd}{Standard deviation}

\item{ll}{Lower bound}

\item{ul}{Upper bound}
}
\description{
Sample from a truncated normal distribution
}
\seealso{
There is a discussion and code on
\url{https://stackoverflow.com/a/14034577}, and there is also the truncnorm
package.
}
\keyword{internal}
