% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_sim_tree.R
\name{run_1sim}
\alias{run_1sim}
\title{Generate IR-Tree Data and Fit Model}
\usage{
run_1sim(gen_model = NULL, fit_model = gen_model, N = NULL,
  sigma = NULL, link = c("probit", "logit"), itempar = NULL,
  backend = c("mirt", "mplus"), dir = NULL, save_rdata = FALSE, ...)
}
\arguments{
\item{gen_model}{A description of the user-specified model that is used to
generate the data. See \code{\link{tree_model}} for more information.}

\item{fit_model}{A description of the user-specified model that is fitted to
the data. See \code{\link{tree_model}} for more information.}

\item{N}{Integer, the number of persons.}

\item{sigma}{Either a matrix or a function that returns a matrix. This matrix
is the variance-covariance matrix of the person parameters that is passed
to \code{\link[MASS]{mvrnorm}}. Note that the order of the person
parameters is taken from the section Processes in the \code{model} (see
\code{\link{tree_model}}).}

\item{link}{String, passed to argument 'LINK' in Mplus. Specifies
the link function.}

\item{itempar}{Either a list or a function that returns a list. The list has
an element \code{beta} and an element \code{alpha}. Each of these is a
matrix of item parameters. Note theat the order of items (rows) is taken from the
section Items and the order of processes (columns) is taken from the
section Processes in the \code{model} (see \code{\link{tree_model}}).}

\item{backend}{String specifying wheter to use mirt or Mplus for
estimation.}

\item{dir}{String, path name where the Mplus files should be
stored and run; the default corresponds to the working directory}

\item{save_rdata}{Logical indicating whether to save the results to an RData
file.}

\item{...}{Other parameters passed to \code{\link{fit_tree_mplus}}.}
}
\value{
List of three elements. \code{fit} contains the Mplus output,
\code{gen_args} contains arguments used to generate the data, and
\code{sim_args} contains arguments used to fit the model.
}
\description{
This function generates data from \code{gen_model}, subsequently fits the
\code{fit_model} in Mplus, and returns the results or saves them to an
external RData file.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link{fit_tree_mplus}}, \code{\link{gen_tree_data}}
}
