% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_sim_tree.R
\name{run_sim}
\alias{run_sim}
\title{Run a Simulation by Generating From and Fitting a Tree Model}
\usage{
run_sim(R = 1, plan = NULL, future_args = list(), gen_model = NULL,
  fit_model = gen_model, N = NULL, sigma = NULL, link = c("probit",
  "logit"), itempar = NULL, dir = NULL, save_rdata = FALSE, ...)
}
\arguments{
\item{R}{Number of replications}

\item{plan}{Parameter passed as argument \code{strategy} to
\code{\link[future]{plan}}. May be set to, for example, \code{multisession}
in order to run the simulations in parallel.}

\item{future_args}{Named list. Parameters passed \code{\link[future]{plan}}.}

\item{gen_model}{A description of the user-specified model that is used to
generate the data. See \code{\link{tree_model}} for more information.}

\item{fit_model}{A description of the user-specified model that is fitted to
the data. See \code{\link{tree_model}} for more information.}

\item{N}{Integer, the number of persons.}

\item{sigma}{Either a matrix or a function that returns a matrix. This matrix
is the variance-covariance matrix of the person parameters that is passed
to \code{\link[MASS]{mvrnorm}}. Note that the order of the person
parameters is taken from the section Processes in the \code{model} (see
\code{\link{tree_model}}).}

\item{link}{String, passed to argument 'LINK' in Mplus. Specifies
the link function.}

\item{itempar}{Either a list or a function that returns a list. The list has
an element \code{beta} and an element \code{alpha}. Each of these is a
matrix of item parameters. Note theat the order of items (rows) is taken from the
section Items and the order of processes (columns) is taken from the
section Processes in the \code{model} (see \code{\link{tree_model}}).}

\item{dir}{String, path name where the Mplus files should be
stored and run; the default corresponds to the working directory}

\item{save_rdata}{Logical indicating whether to save the results to an RData
file.}

\item{...}{Other parameters passed \code{\link{run_1sim}}.}
}
\value{
XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX
}
\description{
The function \code{\link{run_1sim}} generates a data set from an IR-tree
model and fits a possibly different model to that data set. Herein, this
process is repeated \code{R} times, and the argument \code{plan} allows to
run the simulation in parallel.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
