% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sort2}
\alias{sort2}
\title{Sort a Vector Based on a Second Vector}
\usage{
sort2(x = NULL, y = NULL, subset = TRUE)
}
\arguments{
\item{x}{vector that should be sorted}

\item{y}{vector providing the sort order}

\item{subset}{Logical. If \code{TRUE}, \code{x} must be a subset of \code{y}.}
}
\value{
sorted \code{x}
}
\description{
This function takes a vector \code{x} and returns \code{x} in
the order provided in \code{y}.
}
\details{
This function was inspired by an
\href{https://stackoverflow.com/a/2117080}{answer} by \href{https://stackoverflow.com/users/170792}{George Dontas} on Stack Overflow. User
contributions on Stack Overflow are licensed under \href{https://creativecommons.org/licenses/by-sa/4.0/}{CC BY-SA 4.0}.
}
\keyword{internal}
