% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sym_diff}
\alias{sym_diff}
\title{Symmetric difference}
\usage{
sym_diff(x, y)
}
\arguments{
\item{x}{vectors (of the same mode) containing a sequence
    of items (conceptually) with no duplicated values.}

\item{y}{vectors (of the same mode) containing a sequence
    of items (conceptually) with no duplicated values.}
}
\value{
vector
}
\description{
Symmetric difference
}
\details{
This function is based on an
\href{https://stackoverflow.com/a/35949294}{answer} by
\href{https://stackoverflow.com/users/3798973}{sebpardo} on Stack Overflow.
User contributions on Stack Overflow are licensed under \href{https://creativecommons.org/licenses/by-sa/4.0/}{CC BY-SA 4.0}.
}
\seealso{
\code{\link[base:setdiff]{base::setdiff()}}
}
\keyword{internal}
