% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_tree_model.R
\name{tree_model}
\alias{tree_model}
\title{ItemResponseTree Model Syntax}
\usage{
tree_model(model = NULL)
}
\arguments{
\item{model}{Character string with a specific structure as described below.}
}
\value{
List of class \code{tree_model}. It contains the information
extracted from parsing \code{model}.
}
\description{
The ItemResponseTree model syntax describes an IR-tree model. The function
\code{tree_model} turns a user-defined model string into a list that
represents the full model as needed by the package.
}
\section{IRT}{


The \code{model} must contain a section with heading IRT. Therein, the IRT
structure of the model is described in a way resembling the MODEL part of
an Mplus input file. It has a structure of \code{LV BY item@1 item*}, where
\code{LV} is the name of the latent variable/parameter/process, \code{item}
is the name of the observed variable in the data set and this is followed
by the loading. The loading may either be fixed (e.g., to 1) using
\code{@1} or it may be set free using \code{*}.

Each measurement model (i.e., the LV and its items) must appear on a
seperate line ending with a semicolon. Items must be seperated by
whitespace. Linebreaks are allowed.

\preformatted{
  ## Example
  IRT:
  t  BY x1@1, x2*, x3*, x4*, x5*, x6*;
  e  BY x1@1, x2@1, x3@1, x4@1, x5@1, x6@1;
  m  BY x1@1, x2@1, x3@1, x4@1, x5@1, x6@1;}
}

\section{Equations}{


The \code{model} must contain a section with heading Equations if Class is
Tree.
Therein, the model equations are described.
They have a structure similar to \code{Cat = p1*(1-p2)}, where \code{Cat}
is any observed response category in the data set.
The names of the parameters must be equal to those of the latent variables
in the section IRT.

The equations may contain only products and not sums.
That is, it is not possible to estimate genuine mixture models as, for
example, in the package mpt2irt.

Each equation must appear on a separate, non-broken line.

\preformatted{
  ## Example
  Equations:
  1 = (1-m)*(1-t)*e
  2 = (1-m)*(1-t)*(1-e)
  3 = m
  4 = (1-m)*t*(1-e)
  5 = (1-m)*t*e}
}

\section{Subtree}{


The \code{model} may contain a section with heading Subtree.
This is necessary if a process in the model equations (section Equations)
may correspond to different latent variables (section IRT).
For example, when analyzing a Big Five data set, one may wish to specify
only one extremity process e for all items but multiple target traits t,
namely, one for each of the five scales.
In such a case, the section Equations would list only the parameter t,
while the section IRT would list the parameters t1, ..., t5.

In the framework of MPT, one would think of such a situation in terms of
multiple albeit similar trees with specific parameter contraints across
trees.
For example, one would use one tree for each Big Five scale and fix the
response style parameters to equality across trees but not the target trait
parameters.

Each line in this section has a structure of \code{process = subprocess +
  subprocess}, where \code{process} is the name of the process used only in
section Equations and \code{subprocess} it the name of the process used only in
section IRT.
Use one line for each definition.

\preformatted{
  ## Example
  Subtree:
  t = t1 + t2 + t3 + t4 + t5}
}

\section{Class}{


The \code{model} must contain a section with heading Class to specify the
type/class of IRT model to use.
Currently, may be either \code{Tree} or \code{GRM} (graded response model).

\preformatted{
  ## Example
  Class:
  Tree}
}

\section{Addendum}{


The \code{model} may contain a section with heading Addendum if
\code{backend = "mplus"} is used for estimation.
Any code in this section is directly pasted in the \code{MODEL} section of
the Mplus input file.
Use a semicolon at the end of each line; lines must not exceed 90 characters.
Note that the addendum is ignored in \code{\link{gen_tree_data}}.

\preformatted{
  ## Example
  Addendum:
  e WITH t@0;
  m WITH t@0;}
}

\examples{
m1 <- "
# Comment
IRT:
t1 BY x1@1, x2*, x3*;
t2 BY x4@1, x5*, x6*;
e  BY x1@1, x2@1, x3@1, x4@1, x5@1, x6@1;
m  BY x1@1, x2@1, x3@1, x4@1, x5@1, x6@1;

Subtree:
t = t1 + t2

Equations:
1 = (1-m)*(1-t)*e
2 = (1-m)*(1-t)*(1-e)
3 = m
4 = (1-m)*t*(1-e)
5 = (1-m)*t*e

Class:
Tree
"

model <- tree_model(m1)
}
