% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tryCatch.W.E}
\alias{tryCatch.W.E}
\title{tryCatch both warnings (with value) and errors}
\usage{
tryCatch.W.E(expr)
}
\arguments{
\item{expr}{an \R expression to evaluate}
}
\value{
a list with 'value' and 'warning', where
'value' may be an error caught.
}
\description{
Catch \emph{and} save both errors and warnings, and in the case of
a warning, also keep the computed result.
}
\examples{
ItemResponseTrees:::tryCatch.W.E(log(1))
ItemResponseTrees:::tryCatch.W.E(log(-1))
ItemResponseTrees:::tryCatch.W.E(log("a"))
}
\references{
\code{demo(error.catching)}
}
\author{
Martin Maechler;
Copyright (C) 2010-2012  The R Core Team
}
\keyword{internal}
