% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_fit_mirt.R
\name{write_mirt_input}
\alias{write_mirt_input}
\title{Prepare a mirt Model}
\usage{
write_mirt_input(model = NULL, data = NULL)
}
\arguments{
\item{model}{A description of the user-specified model. See
\code{\link{tree_model}} for more information.}

\item{data}{Data frame containing only the items.}
}
\value{
A list with four elements. \code{mirt_string} is the
\code{\link[mirt]{mirt.model}} object; \code{itemtype} and \code{values}
are used as arguments for \code{\link[mirt]{mirt}}; \code{lambda} is the
modified lambda matrix from the \code{model}-argument.
}
\description{
This is an internal function used by \code{\link{fit_tree_mirt}}. It receives its
inputs from the model and the data set and returns a
\code{\link[mirt]{mirt.model}} object.
}
