% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_fit_mplus.R
\name{write_mplus_input}
\alias{write_mplus_input}
\title{Prepare an Mplus Input File}
\usage{
write_mplus_input(model = model, pseudoitems = NULL,
  data_file = NULL, fsco_file = NULL, save_fscores = FALSE,
  integration_points = NULL, estimator = NULL, link = NULL,
  analysis_list = list())
}
\arguments{
\item{model}{A description of the user-specified model. See
\code{\link{tree_model}} for more information.}

\item{pseudoitems}{Data frame as returned from \code{\link{recode_data}}.}

\item{data_file}{Character string naming the full file path of the data set.}

\item{fsco_file}{Character string naming the file used by Mplus to store the factor scores.}

\item{save_fscores}{Logical, wheter to save FSCORES or not.}

\item{integration_points}{Integer, passed to argument 'INTEGRATION' in Mplus.
Number of integration points for numerical integration.}

\item{estimator}{Character string, passed to argument 'ESTIMATOR' in Mplus.}

\item{link}{Character string, passed to argument 'LINK' in Mplus. Specifies
the link function.}

\item{analysis_list}{Named list of strings passed to Mplus' argument
ANALYSIS. For example: \code{analysis_list = list(MITERATIONS = "1000")}.}

\item{lambda}{Matrix as returned from \code{\link{fit_tree_mplus}}.}

\item{addendum}{Character string as returned from \code{\link{tree_model}}.}
}
\value{
A list of of class \code{\link[MplusAutomation]{mplusObject}}
}
\description{
This is an internal function used by \code{\link{fit_tree_mplus}}. It receives its
inputs from the model and the data set and returns an object of class
\code{\link[MplusAutomation]{mplusObject}}.
}
\seealso{
MplusAutomation::mplusObject
}
