% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{augment.irtree_fit}
\alias{augment.irtree_fit}
\title{Augment data with information from an irtree_fit object.}
\usage{
\method{augment}{irtree_fit}(x = NULL, data = NULL, se.fit = TRUE,
  method = "EAP", ...)
}
\arguments{
\item{x}{object of class irtree_fit as returned from  \code{\link[=fit.irtree_model]{fit()}}.}

\item{data}{Optional data frame that is returned together with the predicted
values. Argument is not needed since the data are contained in the fitted
object.}

\item{se.fit}{Logical indicating whether standard errors for the fitted
values should be returned as well.}

\item{method}{type of factor score estimation method. Can be:
\itemize{
    \item \code{"EAP"} for the expected a-posteriori (default)
    \item \code{"MAP"} for the maximum a-posteriori (i.e, Bayes modal)
    \item \code{"ML"} for maximum likelihood
    \item \code{"WLE"} for weighted likelihood estimation
    \item \code{"EAPsum"} for the expected a-posteriori for each sum score
    \item \code{"plausible"} for a single plausible value imputation for each case.
      This is equivalent to setting \code{plausible.draws = 1}
    \item \code{"classify"} for the posteriori classification probabilities (only
      applicable when the input model was of class \code{MixtureClass})
 }}

\item{...}{Additional arguments passed to \code{\link[mirt:fscores]{mirt::fscores()}} if applicable.}
}
\value{
Returns a \link[tibble:tibble-package]{tibble} with one row for each
observation and one (two) additional columns for each latent variable if
\code{se.fit = FALSE} (if \code{se.fit = TRUE}). The names of the new columns start
with \code{.fit} (and \code{.se.fit}).
}
\description{
Augment accepts a model object and a dataset and adds
information about each observation in the dataset, namely, predicted values
in the .fitted column. New columns always begin with a . prefix to avoid
overwriting columns in the original dataset.
}
\details{
Note that \code{method} is used only for objects returned from
\code{irtree_fit(engine = "mirt")}.
}
\examples{
m1 <- "
IRT:
attitude BY Comfort, Work, Future, Benefit;
Class:
GRM
"
model1 <- irtree_model(m1)
data(Science, package = "mirt")

fit1 <- fit(model1, Science, engine = "mirt")

glance(fit1)
tidy(fit1)
augment(fit1)
}
\seealso{
\code{\link[broom:augment]{broom::augment()}}
}
