% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{clps}
\alias{clps}
\title{Wrapper Around Paste to Collapse a Character Vector}
\usage{
clps(collapse = " ", ..., sep = " ")
}
\arguments{
\item{collapse}{an optional character string to separate the results.  Not
    \code{\link[base]{NA_character_}}.}

\item{...}{
  Arguments passed on to \code{\link[base:paste]{base::paste}}
  \describe{
    \item{\code{recycle0}}{\code{\link[base]{logical}} indicating if zero-length
    character arguments should lead to the zero-length
    \code{\link[base]{character}(0)} after the \code{sep}-phase (which turns into
    \code{""} in the \code{collapse}-phase, i.e., when \code{collapse} is
    not \code{NULL}).}
  }}

\item{sep}{a character string to separate the terms.  Not
    \code{\link[base]{NA_character_}}.}
}
\description{
Wrapper Around Paste to Collapse a Character Vector
}
\seealso{
\code{\link[base:paste]{base::paste()}}
}
\keyword{internal}
