% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{clps}
\alias{clps}
\title{Wrapper Around Paste to Collapse a Character Vector}
\usage{
clps(collapse = " ", ..., sep = " ")
}
\arguments{
\item{collapse}{an optional character string to separate the results.  Not
    \code{\link[base]{NA_character_}}.}

\item{...}{
  Arguments passed on to \code{\link[base:paste]{base::paste}}
  \describe{
    \item{\code{}}{}
  }}

\item{sep}{a character string to separate the terms.  Not
    \code{\link[base]{NA_character_}}.}
}
\description{
Wrapper Around Paste to Collapse a Character Vector
}
\seealso{
\code{\link[base:paste]{base::paste()}}
}
\keyword{internal}
