% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit.irtree_model}
\alias{fit.irtree_model}
\title{Fit an IR-Tree model.}
\usage{
\method{fit}{irtree_model}(object = NULL, data = NULL,
  engine = c("mplus", "mirt"), verbose = interactive(), ...)
}
\arguments{
\item{object}{A description of the user-specified model. See
\link{irtree_model} for more information.}

\item{data}{Data frame containing containing one row per respondent and one
column per variable.}

\item{engine}{String specifying whether to use Mplus or mirt for estimation.}

\item{verbose}{Logical indicating whether Mplus/mirt output should be printed
to the console.}

\item{...}{further arguments passed either to \code{\link{irtree_fit_mplus}} or
\code{\link{irtree_fit_mirt}}}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
This function takes a \code{data} frame and a model \code{object} and runs the model in
either Mplus or mirt.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
irtree_fit_mplus irtree_fit_mirt
}
