% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{glance.irtree_fit}
\alias{glance.irtree_fit}
\title{Glance at an irtree_fit object.}
\usage{
\method{glance}{irtree_fit}(x = NULL, ...)
}
\arguments{
\item{x}{object of class \code{irtree_fit} as returned from \code{\link[=fit.irtree_model]{fit()}}.}

\item{...}{Additional arguments. Not used.}
}
\value{
A one-row \link[tibble:tibble-package]{tibble} with columns such as \code{AIC} and \code{BIC}.
}
\description{
Glance accepts an \code{irtree_fit} object and returns a \link[tibble:tibble-package]{tibble}
with exactly one row of model summaries.
}
\section{Converged}{
 The column \code{converged} indicates whether the model
converged or not. For Mplus, this is \code{TRUE} if the output contained the
phrase "The model estimation terminated normally". For mirt, this is equal
to the output of \code{\link[mirt:extract.mirt]{mirt::extract.mirt(x, "converged")}}. You are encoured to check any warnings
or errors in any case.
}

\section{Iterations}{
 \code{iterations} is \code{NA} for Mplus models since respective
information is not easily obtained from the output.
}

\examples{
m1 <- "
IRT:
attitude BY Comfort, Work, Future, Benefit;
Class:
GRM
"
model1 <- irtree_model(m1)
data(Science, package = "mirt")

fit1 <- fit(model1, Science, engine = "mirt", quadpts = 15)

glance(fit1)
tidy(fit1)
augment(fit1)
}
\seealso{
\code{\link[broom:glance]{broom::glance()}}, \code{\link[mirt:extract.mirt]{mirt::extract.mirt(x, "secondordertest")}}
}
