% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irtree_paste.R
\name{irtree_create_template}
\alias{irtree_create_template}
\title{Create a template of a model string}
\usage{
irtree_create_template(variables = NULL, rasch = TRUE)
}
\arguments{
\item{variables}{Character vector of variable names.}

\item{rasch}{Logical. The string \verb{@1} will be appended to each variable name
if \code{TRUE} with no effect otherwise.}
}
\description{
This is a helper function that prints a template of a model string to the
console based on the supplied variable names. This template can be
copy-pasted and modified to define an \code{\link{irtree_model}}.
}
\examples{
irtree_create_template(c("X1", "X2", "X3"))
#> model1 <- "
#> IRT:
#> ... BY X1@1, X2@1, X3@1;
#> ... BY X1@1, X2@1, X3@1;
#>
#> Equations:
#> 0 = ...
#> 1 = ...
#>
#> Class:
#> Tree
#> "
}
\seealso{
irtree_model()
}
