% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irtree_paste.R
\name{irtree_create_template}
\alias{irtree_create_template}
\title{Create a template of a model string}
\usage{
irtree_create_template(data = NULL, rasch = TRUE)
}
\arguments{
\item{data}{Data frame.}

\item{rasch}{Logical. The string \verb{@1} will be appended to each variable name
if \code{TRUE} with no effect otherwise.}
}
\description{
This is a helper function that prints a template of a model string to the
console based on the supplied data frame. This template can be
copy-pasted and modified to define an \link{irtree_model}.
}
\examples{
irtree_create_template(jackson[, c(1, 6, 11)])
#> m1 <- "
#> IRT:
#> ... BY E1@1, E2@1, E3@1;
#> ... BY E1@1, E2@1, E3@1;
#>
#> Equations:
#> 1 = ...
#> 2 = ...
#> 3 = ...
#> 4 = ...
#> 5 = ...
#>
#> Class:
#> Tree
#> "
}
