% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-mirt.R
\name{irtree_fit_mirt}
\alias{irtree_fit_mirt}
\title{Fit an IR-Tree Model using mirt.}
\usage{
irtree_fit_mirt(object = NULL, data = NULL, link = "logit",
  SE = TRUE, verbose = interactive(), rm_mirt_internal = TRUE, ...)
}
\arguments{
\item{object}{A description of the user-specified model. See
\link{irtree_model} for more information.}

\item{data}{Data frame containing containing one row per respondent and one
column per variable. The variable names must correspond to those used in
\code{object}.}

\item{link}{String specifying the link function. Only \code{logit} is
implemented in mirt.}

\item{SE}{logical; estimate the standard errors by computing the parameter information matrix?
See \code{SE.type} for the type of estimates available}

\item{verbose}{Logical indicating whether Mplus/mirt output should be printed
to the console.}

\item{rm_mirt_internal}{Logical. \code{\link[mirt:mirt]{mirt::mirt()}} returns a lot of
information including two functions that can take up a huge amount of space
(https://github.com/philchalmers/mirt/issues/147#issue-352032654). These
two functions are removed from the output if \code{rm_mirt_internal =
  TRUE}.}

\item{...}{Other arguments passed to \code{\link[mirt:mirt]{mirt::mirt()}}.}
}
\value{
List with two elements. \code{mirt} contains the mirt output, namely
an object of class \code{\link[mirt]{SingleGroupClass-class}} . \code{args}
contains the input specifications.
}
\description{
This function takes a \code{data} frame and a model \code{object} and runs the model in mirt.
}
\examples{
m1 <- "
IRT:
attitude BY Comfort, Work, Future, Benefit;
Class:
GRM
"
model1 <- irtree_model(m1)
data(Science, package = "mirt")

fit1 <- fit(model1, Science, engine = "mirt")

glance(fit1)
tidy(fit1)
augment(fit1)
}
