% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-tam.R
\name{irtree_fit_tam}
\alias{irtree_fit_tam}
\title{Fit an IR-Tree Model using TAM.}
\usage{
irtree_fit_tam(
  object = NULL,
  data = NULL,
  link = "logit",
  verbose = interactive(),
  control = control_tam(),
  improper_okay = FALSE
)
}
\arguments{
\item{object}{Object of class \code{irtree_model}. See \link{irtree_model} for more
information.}

\item{data}{Data frame containing containing one row per respondent and one
column per variable. The variable names must correspond to those used in
\code{object}.}

\item{link}{String specifying the link function. Only \code{logit} is
implemented in TAM.}

\item{verbose}{Logical indicating whether output should be printed to the
console.}

\item{control}{List. The allowed elements of this list depend on the
\code{engine}. Use \link{control_mirt}, \link{control_mplus}, or \link{control_tam} for
convenience. Note that the \code{fit()} function does not use \code{...}, but that you
can use the \verb{control_*()} functions to pass additional arguments.}

\item{improper_okay}{Logical indicating whether the model should also be fit
if it is not a proper IR-tree model. Set this only to \code{TRUE} if you really
know what you are doing.}
}
\value{
List with two elements. \code{tam} contains the TAM output, namely
an object of class \code{\link[TAM:tam.mml]{tam.mml}} . \code{spec}
contains the input specifications.
}
\description{
This function takes a \code{data} frame and a model \code{object} and runs the model in TAM.
}
\examples{
m1 <- "
IRT:
attitude BY Comfort, Work, Future, Benefit;
Class:
GRM
"

model1 <- irtree_model(m1)

data(Science, package = "mirt")

fit1 <- fit(model1, data = Science, engine = "mirt",
            control = control_mirt(quadpts = 15))

glance(fit1)
tidy(fit1, par_type = "difficulty")
augment(fit1)
}
