% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-tam.R
\name{irtree_fit_tam}
\alias{irtree_fit_tam}
\title{Fit an IR-Tree Model using TAM.}
\usage{
irtree_fit_tam(object = NULL, data = NULL, link = "logit",
  verbose = interactive(), ..., .set_min_to_0 = FALSE)
}
\arguments{
\item{object}{
Object of class \code{tam} or \code{tam.mml} (only for \code{summary.tam}
functions)
}

\item{data}{Data frame containing containing one row per respondent and one
column per variable. The variable names must correspond to those used in
\code{object}.}

\item{link}{String specifying the link function. Only \code{logit} is
implemented in TAM.}

\item{verbose}{Logical indicating whether output should
be printed during iterations. This argument replaces \code{control$progress}.
}

\item{...}{Other arguments passed to \code{\link[TAM:tam.mml]{TAM::tam.mml()}}.}

\item{.set_min_to_0}{Logical. \link[TAM:tam.mml]{TAM} expects the data to be scored 0,
..., K. If \code{.set_min_to_0 = TRUE}, the minimum of the data is subtracted from
each response, which will likely both satisfy TAM and do no harm to the
data.}
}
\value{
List with two elements. \code{tam} contains the TAM output, namely
an object of class \code{\link[TAM:tam.mml]{tam.mml}} . \code{spec}
contains the input specifications.
}
\description{
This function takes a \code{data} frame and a model \code{object} and runs the model in TAM.
}
\examples{
m1 <- "
IRT:
attitude BY Comfort, Work, Future, Benefit;
Class:
GRM
"
model1 <- irtree_model(m1)
data(Science, package = "mirt")

fit1 <- fit(model1, Science, engine = "mirt")

glance(fit1)
tidy(fit1)
augment(fit1)
}
