% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-pcm-data.R
\name{irtree_gen_pcm}
\alias{irtree_gen_pcm}
\title{Generate Data From a Partial Credit Model}
\usage{
irtree_gen_pcm(object = NULL, N = NULL, sigma = NULL, theta = NULL,
  itempar = NULL, link = "logit", .na_okay = TRUE)
}
\arguments{
\item{object}{Object of class \code{irtree_model}. See \link{irtree_model} for more
information.}

\item{N}{Integer, the number of persons.}

\item{sigma}{Either a matrix or a function that returns a matrix. This matrix
is the variance-covariance matrix of the person parameters that is passed
to \code{\link[MASS:mvrnorm]{MASS::mvrnorm()}}. Note that the order of the person
parameters is taken from the section Processes in the model \code{object} (see
\code{\link[=irtree_model]{irtree_model()}}).}

\item{theta}{Optional numeric matrix of person parameters with one row per person and
one column per dimension (i.e., \code{object$S}). If provided, this overrides
\code{N} and \code{sigma}.}

\item{itempar}{Either a list or a function that returns a list. The list has
an element \code{beta} and an element \code{alpha}. Each of these is a
matrix of item parameters. Note that the order of items (rows) is taken from the
section Items and the order of processes (columns) is taken from the
section Processes in the \code{model} (see \code{\link{irtree_model}}).}

\item{link}{Character. Link function to use.}

\item{.na_okay}{Logical indicating whether variables with unobserved response
categories are permitted. If not (\code{.na_okay = FALSE}), rejection sampling
is used to ensure that all categories are observed.}
}
\value{
A list with element \code{data} containing the data and an
element \code{spec} containing the true parameter values etc.
}
\description{
This function generates data from a (multidimensional) PCM
}
\examples{
m1 <- "
IRT:
t BY x1@1;
e BY x1@1;
m BY x1@1;
Weights:
t = c(0, 1, 2, 3, 4)
e = c(1, 0, 0, 0, 1)
m = c(0, 0, 1, 0, 0)
Class:
PCM
"
model1 <- irtree_model(m1)
dat <- irtree_gen_pcm(model1, N = 5, sigma = diag(3),
                      itempar = list(beta = matrix(rnorm(4), 1, 4)))

m2 <- "
IRT:
t BY x1@1, x2@1, x3@1, x4@1, x5@1;
Weights:
t = c(0, 1, 2, 3, 4)
Class:
PCM
"
model2 <- irtree_model(m2)
dat <- irtree_gen_pcm(model2, N = 5, sigma = diag(1),
                      itempar = list(beta = matrix(sort(rnorm(4*5)), 5, 4)))
}
