% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-data.R
\name{irtree_recode}
\alias{irtree_recode}
\title{Recode Data Set Into Pseudoitems.}
\usage{
irtree_recode(object = NULL, data = NULL, keep = FALSE)
}
\arguments{
\item{object}{A description of the user-specified model. See
\link{irtree_model} for more information.}

\item{data}{Data frame containing containing one row per respondent and one
column per variable.}

\item{keep}{Logical indicating whether to append the original items to the
data frame of the generated pseudoitems}
}
\value{
Data frame
}
\description{
This function takes a data set with polytomous items and an IR-tree model and
returns the recoded items, the so-called pseudoitems.
}
