% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{irtree_sim}
\alias{irtree_sim}
\title{Run a Simulation by Generating From and Fitting a Tree Model}
\usage{
irtree_sim(
  R = 1,
  gen_model = NULL,
  fit_model = gen_model,
  N = NULL,
  sigma = NULL,
  itempar = NULL,
  engine = c("mirt", "mplus", "tam"),
  link = c("logit", "probit"),
  verbose = FALSE,
  control = NULL,
  improper_okay = FALSE,
  save_rdata = TRUE,
  file = NULL,
  plan = NULL,
  plan_args = list(),
  in_memory = c("reduced", "everything", "nothing"),
  difficulty = TRUE,
  dir = tempdir(),
  na_okay = TRUE
)
}
\arguments{
\item{R}{Number of replications. Can be either a single number indicating the
number of replications (e.g., \code{R = 100}), or can be a range (e.g., \code{R = 1:100}).}

\item{gen_model}{Object of class \code{irtree_model} describing the
data-generating model. See \link{irtree_model} for more information.}

\item{fit_model}{Object of class \code{irtree_model} describing the model that
should be fit to the data. May be a list of multiple objects of class
\code{irtree_model} if different models should be fit to the same data set. See
\link{irtree_model} for more information.}

\item{N}{Integer, the number of persons.}

\item{sigma}{Either a matrix or a function that returns a matrix. This matrix
is the variance-covariance matrix of the person parameters that is passed
to \code{\link[MASS:mvrnorm]{MASS::mvrnorm()}}. Note that the order of the person
parameters is taken from the section Processes in the model \code{object} (see
\code{\link[=irtree_model]{irtree_model()}}).}

\item{itempar}{Either a list or a function that returns a list. The list has
an element \code{beta} and an element \code{alpha}. Each of these is a
matrix of item parameters. Note that the order of items (rows) is taken from the
section Items and the order of processes (columns) is taken from the
section Processes in the \code{model} (see \code{\link[=irtree_model]{irtree_model()}}).}

\item{engine}{String specifying whether to use mirt, Mplus, or TAM for
estimation.}

\item{link}{Character. Link function to use.}

\item{verbose}{Logical indicating whether output should be printed to the
console.}

\item{control}{List. The allowed elements of this list depend on the
\code{engine}. Use \link{control_mirt}, \link{control_mplus}, or \link{control_tam} for
convenience. Note that the \code{fit()} function does not use \code{...}, but that you
can use the \verb{control_*()} functions to pass additional arguments.}

\item{improper_okay}{Logical indicating whether the model should also be fit
if it is not a proper IR-tree model. Set this only to \code{TRUE} if you really
know what you are doing.}

\item{save_rdata}{Logical indicating whether to save the results to an RData
file.}

\item{file}{String giving the file path used to save the output if
\code{save_rdata = TRUE}. Note that the file ending is automatically set to
\code{.rda}. This argument is also passed to \code{\link[=irtree_fit_mplus]{irtree_fit_mplus()}} if applicable.}

\item{plan}{Parameter passed as argument \code{strategy} to \code{\link[future:plan]{future::plan()}}. May
be set to, for example, \code{multiprocess} in order to run the simulations in
parallel.}

\item{plan_args}{Named list. Parameters passed \code{\link[future:plan]{future::plan()}}.}

\item{in_memory}{Character string indicating what output should be kept in
memory (note the argument \code{save_rdata}, which is not affected by
\code{in_memory}).}

\item{difficulty}{Logical. The \link[mirt:mirt-package]{mirt} package uses easiness
parameters. These are returned with a message if \code{difficulty = NA} (the
default), without a message if \code{difficulty = FALSE}, and they are
transformed to difficulty parameters if \code{difficulty = TRUE}.}

\item{dir}{Path name that is used to save the results of every run if
\code{save_rdata = TRUE}.}

\item{na_okay}{Logical indicating whether variables with unobserved response
categories are permitted. If not (\code{na_okay = FALSE}), rejection sampling
is used to ensure that all categories are observed.}

\item{...}{Other parameters passed \code{\link[=irtree_sim1]{irtree_sim1()}}.}
}
\value{
Returns a list of length \code{R}, where each element is the output of
\code{\link[=irtree_sim1]{irtree_sim1()}}. If \code{in_memory = "nothing"}, returns \code{NULL}.
}
\description{
The function \code{\link[=irtree_sim1]{irtree_sim1()}} generates a data set from an IR-tree
model and fits one or more models to that data set. Herein, this
process is repeated \code{R} times, and the argument \code{plan} allows to
run the simulation in parallel.
}
\examples{
\donttest{
m1 <- "
IRT:
a BY x1@1, x2@1, x3@1, x4@1, X5@1, X6@1, X7@1, X8@1, X9@1, X10@1;
b BY x1@1, x2@1, x3@1, x4@1, X5@1, X6@1, X7@1, X8@1, X9@1, X10@1;

Equations:
1 = 1-a
2 = a*(1-b)
3 = a*b

Class:
Tree
"

m2 <- "
IRT:
a BY x1@1, x2@1, x3@1, x4@1, X5@1, X6@1, X7@1, X8@1, X9@1, X10@1;

Class:
GRM
"

model1 <- irtree_model(m1)
model2 <- irtree_model(m2)

res1 <- irtree_sim(
    ### Data generation ###
    gen_model = model1,
    N = 500,
    link = "logit",
    sigma = function(x) diag(2),
    itempar = function(x) list(
        beta = matrix(sort(runif(model1$J*model1$P, -2, 2)),
                      model1$J, model1$P),
        alpha = matrix(1, model1$J, model1$P)),
    na_okay = FALSE,

    ### Estimation ###
    fit_model = list(model1, model2),
    engine = "mirt",
    dots = list(fit = list(SE = FALSE),
                tidy = list(difficulty = TRUE)),

    ### Replications ###
    R = 2,
    save_rdata = FALSE,

    ### Parallelization ###
    plan = "multiprocess",
    plan_args = list(workers = 2)
)

library("tidyr")
library("dplyr")
library("tibble")

enframe(res1, "repl", "res") \%>\%
    hoist(res, m1 = list("fits", "m1", "glanced"), m2 = list("fits", "m2", "glanced")) \%>\%
    pivot_longer(cols = c(m1, m2), names_to = "model", values_to = "glanced") \%>\%
    unnest_wider(col = glanced) \%>\%
    select(repl, model, AIC, BIC)
#> # A tibble: 4 x 4
#>   repl  model   AIC   BIC
#>   <chr> <chr> <dbl> <dbl>
#> 1 sim1  m1    9607. 9704.
#> 2 sim1  m2    9773. 9866.
#> 3 sim2  m1    9507. 9603.
#> 4 sim2  m2    9676. 9768.
}
}
\seealso{
\code{\link[=irtree_sim1]{irtree_sim1()}}
}
