% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ItemResponeTrees-package.R
\docType{package}
\name{ItemResponseTrees-package}
\alias{ItemResponseTrees-package}
\alias{_PACKAGE}
\alias{ItemResponseTrees}
\title{Fitting and Simulating IR-Tree Models}
\description{
Item response tree (IR-tree) models are a class of item response models that
reparameterize polytomous responses using a processing tree structure.
They have been used to study response styles but are suitable for other
applications as well.
They can be fit using standard IRT software, and this package uses either the
proprietary software 'Mplus' or the R package 'mirt' as a backend.
Functions are provided to generate data, to fit a model, and to run
simulation studies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hplieninger/ItemResponseTrees}
  \item Report bugs at \url{https://github.com/hplieninger/ItemResponseTrees/issues}
}

}
\author{
Hansjoerg Plieninger
}
