% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ItemResponeTrees-package.R
\docType{package}
\name{ItemResponseTrees-package}
\alias{ItemResponseTrees-package}
\alias{_PACKAGE}
\alias{ItemResponseTrees}
\title{IR-Tree Modeling in mirt, Mplus, or TAM}
\description{
ItemResponseTrees is an R package that allows to fit IR-tree models in Mplus
or mirt. The package automates some of the hassle of IR-tree modeling by
means of a consistent syntax. This allows new users to quickly adopt this
model class, and this allows experienced users to fit many complex models
effortlessly.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hplieninger/ItemResponseTrees}
  \item Report bugs at \url{https://github.com/hplieninger/ItemResponseTrees/issues}
}

}
\author{
Hansjoerg Plieninger
}
