% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{myTryCatch}
\alias{myTryCatch}
\title{Custom tryCatch() that also returns the value of the expression}
\usage{
myTryCatch(expr)
}
\arguments{
\item{expr}{expression to be evaluated}
}
\value{
list with three elements, namely, \code{value} giving the value of
\code{expr}, \code{warning} giving any warning messages, and \code{error}
giving any error messages.
}
\description{
Custom tryCatch() that also returns the value of the expression
}
\examples{
ItemResponseTrees:::myTryCatch(log(1))
ItemResponseTrees:::myTryCatch(log(-1))
ItemResponseTrees:::myTryCatch(log("a"))
}
\references{
https://stackoverflow.com/a/4952908; https://stackoverflow.com/a/24569739
}
\seealso{
\code{\link{tryCatch}}
}
\author{
https://stackoverflow.com/users/2161065
}
\keyword{internal}
