% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sort2}
\alias{sort2}
\title{Sort a Vector Based on a Second Vector}
\usage{
sort2(x = NULL, y = NULL, x_names = FALSE, subset = TRUE)
}
\arguments{
\item{x}{vector that should be sorted}

\item{y}{vector providing the sort order}

\item{x_names}{Logical. If \code{TRUE}, \code{x} is sorted but not on
\code{x} but on \code{names(x)}.}

\item{subset}{Logical. If \code{TRUE}, \code{x} must be a subset of \code{y}.}
}
\value{
sorted \code{x}
}
\description{
This function takes a vector \code{x} and returns \code{x} in
the order provided in \code{y}.
}
\references{
https://stackoverflow.com/a/2117080
}
\author{
George Dontas (https://stackoverflow.com/users/170792)
}
\keyword{internal}
