% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sym_diff}
\alias{sym_diff}
\title{Symmetric difference}
\usage{
sym_diff(x, y)
}
\arguments{
\item{x}{vectors (of the same mode) containing a sequence
    of items (conceptually) with no duplicated values.}

\item{y}{vectors (of the same mode) containing a sequence
    of items (conceptually) with no duplicated values.}
}
\value{
vector
}
\description{
Symmetric difference
}
\references{
https://stackoverflow.com/a/35949294
}
\seealso{
\code{\link[base:setdiff]{base::setdiff()}}
}
\author{
https://stackoverflow.com/users/3798973
}
\keyword{internal}
