% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy.irtree_fit}
\alias{tidy.irtree_fit}
\title{Tidy an irtree_fit object.}
\usage{
\method{tidy}{irtree_fit}(x = NULL, ...)
}
\arguments{
\item{x}{object of class irtree_fit as returned from  \code{\link[=fit.irtree_model]{fit()}}.}

\item{...}{Additional arguments. Not used.}
}
\value{
A \link[tibble:tibble-package]{tibble} with one row for each model
parameter and the following columns:
\describe{
\item{\code{term}}{The name of the model parameter.}
\item{\code{estimate}}{The estimated value of the term.}
\item{\code{std.error}}{The standard error of the term.}
\item{\code{statistic}}{The value of the test statistic of the term (Mplus only).}
\item{\code{p.value}}{The p-value associated with the statistic.}
}
}
\description{
Tidy summarizes information about the parameter estimates of the IR-tree model.
}
\examples{
m1 <- "
IRT:
attitude BY Comfort, Work, Future, Benefit;
Class:
GRM
"
model1 <- irtree_model(m1)
data(Science, package = "mirt")

fit1 <- fit(model1, Science, engine = "mirt")

glance(fit1)
tidy(fit1)
augment(fit1)
}
\seealso{
\code{\link[broom:tidy]{broom::tidy()}}
}
