% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy_mirt}
\alias{tidy_mirt}
\title{Tidy an irtree_fit object estimated using mirt.}
\usage{
tidy_mirt(x = NULL, par_type = NULL)
}
\arguments{
\item{x}{object of class irtree_fit as returned from  \code{\link[=fit.irtree_model]{fit()}}.}

\item{par_type}{Only used if the fit engine was mirt. Item parameters (or
thresholds) can be either of type \code{easiness} (the mirt default) or
\code{difficulty} (as in Mplus and TAM).}
}
\value{
A \link[tibble:tibble-package]{tibble} with one row for each model
parameter and the following columns:
\describe{
\item{\code{term}}{The name of the model parameter.}
\item{\code{estimate}}{The estimated value of the term.}
\item{\code{std.error}}{The standard error of the term.}
\item{\code{statistic}}{The value of the test statistic of the term (Mplus only).}
\item{\code{p.value}}{The p-value associated with the statistic.}
}
}
\description{
Tidy summarizes information about the parameter estimates of the IR-tree model.
}
\seealso{
\code{\link[=tidy.irtree_fit]{tidy()}}, \code{\link[generics:tidy]{generics::tidy()}}
}
