% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-mirt.R
\name{write_mirt_input}
\alias{write_mirt_input}
\title{Prepare a mirt Model}
\usage{
write_mirt_input(object = NULL, data = NULL)
}
\arguments{
\item{object}{Object of class \code{irtree_model}. See \link{irtree_model} for more
information.}

\item{data}{Data frame containing containing one row per respondent and one
column per variable. The variable names must correspond to those used in
\code{object}.}
}
\value{
A list with four elements. \code{mirt_string} is the \link[mirt:mirt.model]{mirt::mirt.model}
object; \code{itemtype} and \code{values} are used as arguments for \code{\link[mirt:mirt]{mirt::mirt()}};
\code{lambda} is the modified lambda matrix from the \code{object}-argument.
}
\description{
This is an internal function used by \code{\link[=irtree_fit_mirt]{irtree_fit_mirt()}}. It receives its
inputs from the model object and the data set and returns a
\link[mirt:mirt.model]{mirt::mirt.model} object.
}
