% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-mplus.R
\name{write_mplus_input}
\alias{write_mplus_input}
\title{Prepare an Mplus Input File}
\usage{
write_mplus_input(object = object, pseudoitems = NULL,
  data_file = NULL, fsco_file = NULL, save_fscores = FALSE,
  integration_points = NULL, estimator = NULL, link = NULL,
  analysis_list = list())
}
\arguments{
\item{object}{Object of class \code{irtree_model}. See \link{irtree_model} for more
information.}

\item{pseudoitems}{Data frame as returned from \code{\link{irtree_recode}}.}

\item{data_file}{String, the full file path of the data set.}

\item{fsco_file}{String, the file name used by Mplus to store the factor scores.}

\item{save_fscores}{Logical, wheter to save FSCORES or not.}

\item{integration_points}{Integer, passed to argument 'INTEGRATION' in Mplus.
Number of integration points for numerical integration.}

\item{estimator}{String, passed to argument 'ESTIMATOR' in Mplus.}

\item{link}{String, passed to argument 'LINK' in Mplus. Specifies
the link function.}

\item{analysis_list}{Named list of strings passed to Mplus' argument
ANALYSIS. For example: \code{analysis_list = list(MITERATIONS = "1000")}.}
}
\value{
A list of of class \link[MplusAutomation:mplusObject]{MplusAutomation::mplusObject}
}
\description{
This is an internal function used by \code{\link[=irtree_fit_mplus]{irtree_fit_mplus()}}. It receives its
inputs from the model object and the data set and returns an object of class
\link[MplusAutomation:mplusObject]{MplusAutomation::mplusObject}.
}
\seealso{
\link[MplusAutomation:mplusObject]{MplusAutomation::mplusObject}
}
